requirements.txt
streamlit
python-dotenv
eventregistry
wordcloud
matplotlib
langchain
openai
streamlit-agraph
faiss-cpu
streamlit-tags
tiktoken